[trello-106]: Moved userpanel tests into smaller files
